requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
psycopg2-binary
python-jose[cryptography]
passlib
pydantic[email]
bcrypt==4.0.1